Refactor cli stuff out to a separate file
