+++ package.json
@@ -56,6 +56,7 @@
     "@types/jest": "^25.2.1",
     "@types/js-yaml": "^3.12.1",
     "@types/node": "^13.11.1",
+    "@types/node-fetch": "^2.5.6",
     "@types/supertest": "^2.0.8",
     "@types/yup": "^0.26.26",
     "@typescript-eslint/eslint-plugin": "^2.7.0",
